+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/annotation/VariantDescription.java
@@ -1,0 +1,10 @@
+package de.charite.compbio.jannovar.annotation;
+
+/**
+ * Superclass of both {@link SmallVariantDescription} and {@link 
+ * 
+ * @author Manuel Holtgrewe <manuel.holtgrewe@bihealth.de>
+ */
+public class VariantDescription {
+
+}